% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{analyze_concordance_and_markers}
\alias{analyze_concordance_and_markers}
\title{Esegue l'analisi di concordanza e trova i geni marcatori.}
\usage{
analyze_concordance_and_markers(scrna_obj, group_by_labels = "singleR_labels")
}
\arguments{
\item{scrna_obj}{Un oggetto Seurat completo di clustering e annotazione.}

\item{group_by_labels}{La colonna dei metadati da usare per trovare i marcatori (es. "singleR_labels").}
}
\value{
Un data.frame con i geni marcatori per ogni gruppo.
}
\description{
Calcola e stampa una tabella di concordanza tra i cluster di Seurat
e le annotazioni di SingleR. Successivamente, identifica i geni marcatori
positivi per ogni tipo cellulare annotato.
}
