% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{annotate_cell_types}
\alias{annotate_cell_types}
\title{Annota i tipi cellulari con SingleR.}
\usage{
annotate_cell_types(scrna_obj)
}
\arguments{
\item{scrna_obj}{Un oggetto Seurat.}
}
\value{
L'oggetto Seurat con una nuova colonna nei metadati ('singleR_labels')
contenente le annotazioni.
}
\description{
Usa il pacchetto SingleR e il riferimento HumanPrimaryCellAtlasData
per predire i tipi cellulari per ogni cellula.
}
