% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{generate_visualizations}
\alias{generate_visualizations}
\title{Crea e visualizza i grafici di analisi principali.}
\usage{
generate_visualizations(scrna_obj)
}
\arguments{
\item{scrna_obj}{Un oggetto Seurat completo di PCA, clustering e annotazione.}
}
\value{
Una lista di oggetti ggplot (i grafici).
}
\description{
Genera una serie di grafici standard per visualizzare i risultati
dell'analisi: la varianza della PCA, la UMAP colorata per cluster e la UMAP
colorata per annotazione cellulare.
}
