% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{load_and_filter_data}
\alias{load_and_filter_data}
\title{Carica i dati 10X e filtra per geni codificanti proteine.}
\usage{
load_and_filter_data(data_dir, gtf_path)
}
\arguments{
\item{data_dir}{Percorso della cartella contenente i file 'matrix.mtx.gz',
'barcodes.tsv.gz', e 'features.tsv.gz'.}

\item{gtf_path}{Percorso del file di annotazione GTF (es. 'Homo_sapiens.GRCh38.111.gtf').}
}
\value{
Un oggetto Seurat contenente solo i geni codificanti proteine.
}
\description{
Questa funzione legge i dati da una cartella 10X, crea un oggetto Seurat,
e lo filtra per mantenere esclusivamente i geni annotati come 'protein_coding'
da un file GTF.
}
