% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{preprocess_and_run_pca}
\alias{preprocess_and_run_pca}
\title{Esegue normalizzazione, scaling e PCA.}
\usage{
preprocess_and_run_pca(scrna_obj, n_variable_features = 2000)
}
\arguments{
\item{scrna_obj}{Un oggetto Seurat.}

\item{n_variable_features}{Numero di geni altamente variabili da identificare. Default 2000.}
}
\value{
L'oggetto Seurat processato e con i risultati della PCA.
}
\description{
Esegue i passi standard di pre-processing di Seurat (Normalize,
FindVariableFeatures, ScaleData) e infine esegue la PCA.
}
