% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{run_umap_and_cluster}
\alias{run_umap_and_cluster}
\title{Esegue UMAP e clustering.}
\usage{
run_umap_and_cluster(scrna_obj, pca_dims, resolution = 0.5)
}
\arguments{
\item{scrna_obj}{Un oggetto Seurat con PCA già calcolata.}

\item{pca_dims}{Il numero di PC da usare per l'analisi.}

\item{resolution}{La risoluzione per l'algoritmo di clustering. Valori più alti
portano a più cluster. Default 0.5.}
}
\value{
L'oggetto Seurat con i risultati di UMAP e clustering.
}
\description{
Esegue l'algoritmo UMAP e il clustering basato su grafo (FindNeighbors, FindClusters)
usando un numero definito di componenti principali.
}
